refactor: split cancelBookingModal
